Refactored to new package and class
