+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/DocumentGenerationController.java
@@ -15,8 +15,8 @@ import uk.gov.hmcts.ecm.common.model.ccd.CCDRequest;
 import uk.gov.hmcts.ecm.common.model.ccd.CaseDetails;
 import uk.gov.hmcts.ecm.common.model.helper.DefaultValues;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.Helper;
-import uk.gov.hmcts.ethos.replacement.docmosis.helpers.LettersHelper;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.dynamiclists.DynamicLetters;
+import uk.gov.hmcts.ethos.replacement.docmosis.helpers.letters.InvalidCharacterCheck;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.DefaultValuesReaderService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.DocumentGenerationService;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.EventValidationService;
@@ -225,7 +225,7 @@ public class DocumentGenerationController {
 
         var caseData = ccdRequest.getCaseDetails().getCaseData();
         DynamicLetters.dynamicLetters(caseData, ccdRequest.getCaseDetails().getCaseTypeId());
-        List<String> errors = LettersHelper.checkNamesForInvalidCharacters(caseData);
+        List<String> errors = InvalidCharacterCheck.checkNamesForInvalidCharacters(caseData);
         return getCallbackRespEntityErrors(errors, caseData);
     }
 }
